+++ CHANGELOG.md
@@ -14,6 +14,7 @@ HEAD
 * Fix `JsonVariant::memoryUsage()` for raw strings
 * Fix `call of overloaded 'swap(BasicJsonDocument&, BasicJsonDocument&)' is ambiguous` (issue #1678)
 * Fix inconsistent pool size in `BasicJsonDocument`'s copy constructor
+* Support NUL in string values (issue #1646)
 
 v6.18.5 (2021-09-28)
 -------